requirements.txt
python-dotenv==1.0.0
openai==1.3.5
langchain==0.0.340
pypdf==3.17.1
chromadb==0.4.18
tiktoken==0.5.1